+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
   the optional `signingBox` as an alternative to the `keyPair`.
 - `getCryptoBox` method of the `contracts` module. It creates a default `TONCryptoBox` implementation 
    that uses the core crypto module.
- 
+- More diagnostic fields `configServer`, `queryUrl` in error object. 
 
 ## 0.25.5 - August 5, 2020
 ### Fixed